refactor(db-schema): file-organization sweep over services and the package root
The pgTable overload set is a model file, and the AchievementNames set leaves the enum file: its one reader was the exhaustiveness test, which re-derives it — the case the typescript skill's values-array rule names as the test-side exception.
