+++ .agents/ledgers/file-organization.md
@@ -29,7 +29,7 @@ The question is where a thing lives and whether it exists twice — one export p
 | `app/components` — the rest                                                | —          | a `*Props.ts` beside its component is colocation, not a stray model                          |
 | `packages/db-schema` — `models`                                            | —          |                                                                                              |
 | `packages/db-schema` — `schema`, `relations`                               | —          | a table file declares the table and its select schema; enums live in `models`                |
-| `packages/db-schema` — `services`, root                                    | —          |                                                                                              |
+| `packages/db-schema` — `services`, root                                    | 2026-09-12 |                                                                                              |
 | `packages/db` — `services/azure`                                           | 2026-09-12 |                                                                                              |
 | `packages/db` — the rest, `packages/db-mock`                               | 2026-09-12 |                                                                                              |
 | `packages/azure`, `apps/functions`, `packages/azure-mock`                  | —          | cross-package placement: an Azure helper two packages need lives in `db`                     |
